Implement Runtime wrapper
Using this wrapper, it is possible obtain a specifier runtime environment
with valid pre-defined environment variables.

Each runtime can spawn individual background process and run command as
if it was on the command line. Each runtime also save all stdout and
stderr for further inspection and validation.

Include fix-up to Project objects to accomodate the Runtime object
requirement.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>